requirements.txt
attrs==20.3.0
boto3==1.17.44
botocore==1.20.44
cached-property==1.5.2
click==7.1.2
importlib-metadata==3.10.0
iniconfig==1.1.1
jmespath==0.10.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.3.6
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
